refactor(plugins): extract plugin import scanner to whiskin build module
P2 slice 2. Move validatePluginImports (+ its helpers, the host-provided set,
builtins, and PluginPackageJson) out of user-plugin-builder.ts into
src/core/whiskin/import-scan.ts, so the shared build backend (system-bun build +
publish path) and the in-process dev build use one scanner, with the
host-provided set sourced from the single externals contract.

- The scanner chain was fully self-contained; pure move, no logic change.
- NON_RUNTIME_DIRS is exported (also used by the builder's mtime/staleness walk);
  the now-unused builtinModules import is dropped.

Behavior-preserving: typecheck clean; build-path tests 37/37; build:plugins
builds all 10 core plugins.
